+++ src/components/utils/salesOrderApi.js
@@ -1,11 +1,20 @@
 import { api, getHeader, getToken, getHeaderForFormData } from "./AppFunction";
 import moment from "moment";
 
+const isOrderStatusValid = ["CREATED","PAID","SHIPPED","CANCELLED","PENDING"];
+
 export async function createSalesOrder(buyerId, items, orderDate, totalAmount, note, status, invoiceId, orderNumber){
     try{
-        if (!buyerId || items.length === 0 || !orderDate || !totalAmount || !status || !orderNumber) {
-            throw new Error("Sva obavezna polja moraju biti popunjena.");
-}       
+        if(
+            !buyerId ||!Array.isArray(items) || items.length === 0 ||
+            !moment(orderDate, moment.ISO_8601, true).isValid() ||
+            isNaN(totalAmount) || parseFloat(totalAmount) <= 0 ||
+            !note || typeof note !== "string" || note.trim() ===""||
+            !isOrderStatusValid.includes(status.toUpperCase()) ||
+            !invoiceId || !orderNumber || typeof orderNumber !=="string" || orderNumber.trim() === ""
+        ){
+            throw new Error("Sva polja moraju biti validna i popunjena");
+        }  
         const requestBody = {buyerId, items, orderDate:moment(orderDate).toISOString(),
             totalAmount:parseFloat(totalAmount), note:note, status: status.toUpperCase(), invoiceId:invoiceId, orderNumber:orderNumber
         };
@@ -26,6 +35,17 @@ export async function createSalesOrder(buyerId, items, orderDate, totalAmount, n
 
 export async function updateSalesOrder(id,buyerId, items, orderDate, totalAmount, note, status, invoiceId, orderNumber ){
     try{
+        if(
+            !id ||
+            !buyerId ||!Array.isArray(items) || items.length === 0 ||
+            !moment(orderDate, moment.ISO_8601, true).isValid() ||
+            isNaN(totalAmount) || parseFloat(totalAmount) <= 0 ||
+            !note || typeof note !== "string" || note.trim() ===""||
+            !isOrderStatusValid.includes(status.toUpperCase()) ||
+            !invoiceId || !orderNumber || typeof orderNumber !=="string" || orderNumber.trim() === ""
+        ){
+            throw new Error("Sva polja moraju biti validna i popunjena");
+        }
         const requestBody = {buyerId,items, orderDate:moment(orderDate).toISOString(),totalAmount:parseFloat(totalAmount),  note,
             status: status.toUpperCase(), invoiceId, orderNumber
         };
@@ -46,6 +66,9 @@ export async function updateSalesOrder(id,buyerId, items, orderDate, totalAmount
 
 export async function deleteSalesOrder(id){
     try{
+        if(!id){
+            throw new Error("Dati ID nije pronadjen");
+        }
         const response = await api.delete(`${import.meta.env.VITE_API_BASE_URL}/salesOrders/delete/${id}`,{
             headers:getHeader()
         });
@@ -58,6 +81,9 @@ export async function deleteSalesOrder(id){
 
 export async function getSalesOrderById(id){
     try{
+        if(!id){
+            throw new Error("Dati ID nije pronadjen");
+        }
         const response = await api.get(`${import.meta.env.VITE_API_BASE_URL}/salesOrders/get-one/${id}`,{
             headers:getHeader()
         });
