+++ src/components/utils/shiftApi.js
@@ -1,9 +1,15 @@
 import { api, getHeader, getToken, getHeaderForFormData } from "./AppFunction";
 import moment from "moment";
 
-
 export async function createShift(startTime, endTime, shiftSupervisorId){
     try{
+        if(
+            !moment(startTime, moment.ISO_8601, true).isValid() ||
+            !moment(endTime, moment.ISO_8601, true).isValid() ||
+            !shiftSupervisorId
+        ){
+            throw new Error("Sva polja moraju biti validna i popunjena");
+        }
         const requestBody = {
             startTime:moment(startTime).toISOString(), endTime:moment(endTime).toISOString(),
             shiftSupervisorId: shiftSupervisorId
@@ -25,6 +31,14 @@ export async function createShift(startTime, endTime, shiftSupervisorId){
 
 export async function updateShift(id,startTime, endTime, shiftSupervisorId){
     try{
+        if(
+            !id ||
+            !moment(startTime, moment.ISO_8601, true).isValid() ||
+            !moment(endTime, moment.ISO_8601, true).isValid() ||
+            !shiftSupervisorId
+        ){
+            throw new Error("Sva polja moraju biti validna i popunjena");
+        }
         const requestBody = {id:id, startTime:moment(startTime).toISOString(), endTime:moment(endTime).toISOString(),
             shiftSupervisorId:shiftSupervisorId
         };
@@ -45,6 +59,9 @@ export async function updateShift(id,startTime, endTime, shiftSupervisorId){
 
 export async function deleteShift(id){
     try{
+        if(!id){
+            throw new Error("Dati ID nije nadjen");
+        }
         const response = await api.delete(`${import.meta.env.VITE_API_BASE_URL}/shifts/delete/${id}`,{
             headers:getHeader()
         });
@@ -56,7 +73,10 @@ export async function deleteShift(id){
 }
 
 export async function getShiftById(id){
-    try{    
+    try{   
+        if(!id){
+            throw new Error("Dati ID nije nadjen");
+        } 
         const response = await api.get(`${import.meta.env.VITE_API_BASE_URL}/shifts/get-one/${id}`,{
             headers:getHeader()
         });
